[#7] Refactor : Entity 연관관계 제거
- 빠른 CRUD 기능 구현을 위해, 우선 연관관계 제거
- 기능 구현 후 추후 매핑 예정
